pyproject.toml
[tool.poetry]
name = "space-api"
version = "0.1.0"
description = ""
authors = ["valhallajazzy <valhallajazzy@mail.ru>"]
readme = "README.md"
packages = [{include = "space_api"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-telegram-bot = "13.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
